requirements.txt
fastapi

absl-py==2.1.0
albucore==0.0.17
albumentations==1.4.18
appdirs==1.4.4
astunparse==1.6.3
atomicwrites==1.4.0
beautifulsoup4==4.12.2
boto3
sounddevice
soundfile
numpy
uvicorn[standard]
CacheControl==0.14.1
Cython==3.0.11
eval_type_backport==0.2.0
firebase-admin==6.6.0
flatbuffers==24.3.25
gast==0.6.0
git-filter-repo==2.45.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.71.0
grpcio-status==1.68.1
grpcio-tools==1.71.0
httplib2==0.22.0
huggingface-hub==0.26.5
jsonpointer==2.1
keras==3.6.0
libclang==18.1.1
ml-dtypes==0.4.1
mysql-connector-python==9.1.0
namex==0.0.8
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
pandas==2.1.3
proto-plus==1.25.0
protobuf==5.29.1
pyasn1-modules==0.2.8
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pyls-spyder==0.4.0
PyMySQL==1.1.1
PyQt5==5.15.10
PyQtWebEngine==5.15.6
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
websockets

s3transfer==0.10.4
safetensors==0.4.5
setuptools==75.6.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tk==0.1.0
tokenizers==0.21.0
torch==2.5.1
transformers==4.47.0
uritemplate==4.1.1
utils==1.0.2
wheel==0.45.1
